Remap user ids on restore
I have not checked that this leaves backup in a sensible state!  The
"local_user" setting for the backup step needs to be the user that
owns the starport directory.

This is basically what we had before.  The difference is that we
are passing in the uid (and assuming same gid) rather than getting
them from the docker container process which was running with uid=0.

The experimental approach was failing because the user associated
with uid 999 was not existing for ssh - rsync does not care (I've
tested creating with uid 1009 too and that works).  So we don't need
to do any useradd craziness.

Running as a different user also caused big problems with getting
the ssh configuration and apparently superuser is also needed for
doing the chown bits.